!st milestone

+++ views/AddQuantity.js
@@ -15,18 +15,23 @@
 	import {ListItem, Avatar, Header, Button, Input} from 'react-native-elements';
 	import MainScreen from '../layout/MainScreen';
 	import {useState, useEffect} from 'react';
-	import {getCartItemDetails, getVehicle, imagePrefix} from '../api/apiService';
+	import {getCartItemDetails, getVehicle, imagePrefix, showToast} from '../api/apiService';
 	import AsyncStorage from '@react-native-async-storage/async-storage';
 	import {SaveOrder} from '../api/apiService';
-	import { ActivityIndicator ,Modal} from 'react-native';
+	import { ActivityIndicator ,Modal,Keyboard} from 'react-native';
 	import { useRef } from 'react';
-
+	import { CheckBox } from 'react-native-elements'
+	
 	const win = Dimensions.get('window');
 
 	let setTotalAmount = 0;
+	let totalAmountVatWithout = 0;
+	let totalAmountVat = 0;
 	let setUpdatedDataArray = [];
 	let currentSelectedId = '';
+	let VATUpdatedStatus = [];
 	let currentSelectedLoadName = '';
+	let Vsta = '';
 	let selectedVehicle = '';
 	let selectedRoute = '';
 	let selectedDriver = '';
@@ -36,44 +41,67 @@
 	let updatedValue= '';
 	let initalPaymentStatus = 'cash';
 	export default function AddQuantity({navigation}) {
-		const [data, setData] = useState();
-		// const [totalAmount, setTotalAmount] = useState();
+		const [data, setData] = useState({});
+		// const [totalAmount, setTotalAmount] = useState(0);
 		const [loadedData , setLoadedData] = useState();
 		const [updatedData , setUpdatedData] = useState();
+		const [creaditStatus , setCreditStatus] = useState(initalPaymentStatus);
 		const [loadedActivityIndicator , setLoadedActivityIndicator] = useState(false);
 		const [ActInd , setActInd] = useState(false);
-		const [creaditStatus , setCreditStatus] = useState(initalPaymentStatus);
 		const [saveOrderActivIndictor , setSaveOrderActivIndictor] = useState(false);
 		const [vatStatu , setVATstatus] = useState(false);
 		const [ hasNonVatProducts ,setHasNonVatProducts ] = useState(false);
 		const [ hasVatProducts ,setHasVatProducts ] = useState(false);
-		const ref_input2 = useRef();
+	  	const [toggleCheckBox, setToggleCheckBox] = useState(false)
+	  	const [vatStatusForProducts, setVATstatusForProducts] = useState()
+	  	const [IsKeyboardOpen, setIsKeyboardOpen] = useState(false)
+	  	const [modalVisible, setModalVisible] = useState(false);
+	  	const [MyTotalPrice, setMyTotalPrice] = useState(0);
+		const [hasUndeliveredItems , setHasUndeliveredItems] = useState(true);
+
+		let undeliveredItems = {"LOAD-20-03-2021-246__134":{"value":5,"cardId":134,"VATstatus":false},"LOAD-20-03-2021-246__311":{"value":6,"cardId":311,"VATstatus":false}};
+
+
+		  const ref_input2 = useRef();
+
 
 		useEffect(() => {
+			totalAmountVatWithout = 0;
+			totalAmountVat = 0;
 			selectedLoadedItemsByQty();
 			
+			// return () => { 
+				// 	// AsyncStorage.removeItem('finalItems');
+				// 	setUpdatedDataArray = [];
+				// 	// AsyncStorage.setItem('finalItems' , JSON.stringify({}))
+				// }
+				return () => {
+					totalAmountVatWithout = 0;
+					totalAmountVat = 0;
+					setUpdatedDataArray = [];
+			}
 		} , [])
 		
 		function selectedLoadedItemsByQty() {
+			AsyncStorage.setItem('undeliveredItems' , JSON.stringify(undeliveredItems));
+
 			setLoadedActivityIndicator(true)
 			
 			AsyncStorage.getItem('VATStatus').then((data) => {
-				console.log(typeof data);
 				if( data == 'true' ){
 					AsyncStorage.setItem('currentVATstatus' , '1');
 				}else{
 					AsyncStorage.setItem('currentVATstatus' , '0');
 				}
-
+				
 				setVATstatus(data);
 			});
-
+			
 			AsyncStorage.getItem('selectedLoadedItemsByQty').then((data) => {
 				setLoadedData(JSON.parse(data));
 				getCartItemDetails(data).then((res) => {
 					let productData = res.data.data;
-					console.log( Object.values(productData).length );
-
+					
 					for(let i = 0; i < productData.length ; i++){
 						let myData = Object.values(productData)[i]; 
 
@@ -85,6 +113,17 @@
 					}
 
 					setData(productData);
+					let price = 0;
+					for(let i = 0; i < productData.length ; i++){
+						let myData = Object.values(productData)[i];
+						let myNewData = Object.values(myData)[0]
+						if( myNewData['VATstatus'] == true ){
+							price = price + ((myNewData['sale_price'] * myNewData['order_qty'])*1.20) ;
+						}else{
+							price = price + (myNewData['sale_price'] * myNewData['order_qty'])  ;
+						}
+					}
+					setMyTotalPrice(price);
 					setLoadedActivityIndicator(false)
 				});
 			});
@@ -101,6 +140,11 @@
 					});
 				});
 			});
+
+			// AsyncStorage.getItem('itemsForVAT').then((res) => {
+			// 	let VATDataRecieved = JSON.parse(res);
+			// 	setVATstatusForProducts( VATDataRecieved);
+			// })
 		}
 
 		function generateRandString(){
@@ -129,28 +173,36 @@
 				resolve(Object.values(processedData))
 			})
 		}
-		
-		function SaveOrders(){
-			setSaveOrderActivIndictor(true)
-			AsyncStorage.setItem('finalItems' , JSON.stringify(setUpdatedDataArray));
-			
-			updateRecords(setUpdatedDataArray).then((res) => {
-				let data = [];
-				data.push(res)
-				data.push({'type' : creaditStatus});
-				AsyncStorage.getItem('currentVATstatus').then((VATstatus) => {
-					data.push({'has_vat' : parseInt(VATstatus)});
-					
-					SaveOrder(JSON.stringify(data)).then((res) => {
-						setSaveOrderActivIndictor(false)
-						AsyncStorage.setItem('orderSaveReponce', JSON.stringify(res.data.data));
-						
-						AsyncStorage.setItem('orderSaveBuyer', JSON.stringify(res.data.buyer));
-						navigation.navigate('PDFmanager');
-					})
-				})
 
-			})
+		function showConfirmationModel(){
+			setModalVisible(true);
+		}
+
+		function SaveOrders(){	
+			setModalVisible(!modalVisible)		
+			if( IsKeyboardOpen){
+				Keyboard.dismiss();
+			}else{
+				setSaveOrderActivIndictor(true)
+				AsyncStorage.setItem('finalItems' , JSON.stringify(setUpdatedDataArray));
+				
+				updateRecords(setUpdatedDataArray).then((res) => {
+					let data = [];
+					data.push(res)
+					data.push({'type' : creaditStatus});
+					AsyncStorage.getItem('currentVATstatus').then((VATstatus) => {
+						data.push({'has_vat' : parseInt(VATstatus)});
+						SaveOrder(JSON.stringify(data)).then((res) => {
+							setSaveOrderActivIndictor(false)
+							AsyncStorage.setItem('orderSaveReponce', JSON.stringify(res.data.data));
+							AsyncStorage.setItem('orderSaveBuyer', JSON.stringify(res.data.buyer));
+							showToast('Order has been placed successfully')
+							navigation.navigate('PDFmanager');
+						})
+					})
+	
+				})
+			}
 
 		}
 
@@ -172,7 +224,54 @@
 			})
 		}
 
+		function updateVATStatusOfProduct(dnum , itemId){
+			totalAmountVatWithout = 0;
+			totalAmountVat = 0;
+
+			let myData = Object.values(data);		
+			AsyncStorage.getItem('selectedLoadedItemsByQty').then((res) => {
+				let objectData = JSON.parse(res);
+				if( dnum+'__'+itemId in objectData){
+					if(objectData[dnum+'__'+itemId]['VATstatus'] == false){
+						objectData[dnum+'__'+itemId]['VATstatus'] = true;
+					}else{
+						objectData[dnum+'__'+itemId]['VATstatus'] = false;
+					}
+
+					if(VATUpdatedStatus.includes(dnum+'__'+itemId)){
+						VATUpdatedStatus.pop(dnum+'__'+itemId);
+					}else{
+						VATUpdatedStatus.push(dnum+'__'+itemId);
+					}
+
+					AsyncStorage.setItem('selectedLoadedItemsByQty' ,JSON.stringify(objectData));
+					AsyncStorage.setItem('itemsForVAT' , JSON.stringify(VATUpdatedStatus))
+					setVATstatusForProducts( VATUpdatedStatus);
+				}
+			});
+			for( let i= 0 ; i < myData.length; i++ ){
+				if( myData[i][dnum].id == itemId){
+					if(myData[i][dnum].VATstatus == false){
+						myData[i][dnum].VATstatus = true;
+					}else{
+						myData[i][dnum].VATstatus = false;
+					}
+					setData(myData);
+				}
+			}
+			let price = 0;
+			for( let i= 0 ; i < myData.length; i++ ){
+				if( myData[i][dnum]['VATstatus'] == true ){
+					price = price + ((myData[i][dnum]['sale_price'] * myData[i][dnum]['order_qty'])*1.20) ;
+				}else{
+					price = price + (myData[i][dnum]['sale_price'] * myData[i][dnum]['order_qty'])  ;
+				}
+			}
+			setMyTotalPrice(price);
+		}
+
 		function updateQty(dnum , itemId , qty ){
+			console.log(dnum , itemId , qty);
 			let myData = Object.values(data);
 			let newQty = 0;
 			
@@ -189,10 +288,19 @@
 			for( let i= 0 ; i < myData.length; i++ ){
 				if( myData[i][dnum].id == itemId){
 					myData[i][dnum].order_qty = newQty;
-					
 					setData(myData)
 				}
 			}
+			let price = 0;
+			for( let i= 0 ; i < myData.length; i++ ){
+				if( myData[i][dnum]['VATstatus'] == true ){
+					price = price + ((myData[i][dnum]['sale_price'] * myData[i][dnum]['order_qty'])*1.20) ;
+				}else{
+					price = price + (myData[i][dnum]['sale_price'] * myData[i][dnum]['order_qty'])  ;
+				}
+			}
+			setMyTotalPrice(price);
+			setIsKeyboardOpen(false)
 		}
 
 		function updatePrice(dnum , itemId , value ){
@@ -216,6 +324,16 @@
 					setData(myData)
 				}
 			}
+			let price = 0;
+			for( let i= 0 ; i < myData.length; i++ ){
+				if( myData[i][dnum]['VATstatus'] == true ){
+					price = price + ((myData[i][dnum]['sale_price'] * myData[i][dnum]['order_qty'])*1.20) ;
+				}else{
+					price = price + (myData[i][dnum]['sale_price'] * myData[i][dnum]['order_qty'])  ;
+				}
+			}
+			setMyTotalPrice(price);
+			setIsKeyboardOpen(false)
 			
 		}
 
@@ -235,14 +353,44 @@
 		return (
 			<MainScreen>
 				<View style={{flex:1}}>
-					{( vatStatu != 'true'  && hasVatProducts) ?
+					<Modal
+						animationType="slide"
+						transparent={true}
+						visible={modalVisible}
+						onRequestClose={() => {
+							setModalVisible(!modalVisible);
+						}}
+					>
+						<View style={styles.centeredView}>
+							<View style={styles.modalView}>
+								<Text style={{fontSize: 17,paddingHorizontal: 10}}>You are about to place the order, you will not able to edit this order.</Text>
+								<View style={{flexDirection: 'row',width : '90%',marginTop: 10 ,justifyContent: 'space-between'}}>
+									<Pressable
+										style={[styles.button, styles.buttonClose]}
+										onPress={() => setModalVisible(!modalVisible)}
+									>
+										<Text style={{backgroundColor: 'red' , borderRadius: 2,color: 'white',paddingVertical: 10 ,paddingHorizontal: 13,elevation: 5}}>Cancel</Text>
+									</Pressable>
+									<Pressable
+										style={[styles.button, styles.buttonClose]}
+										onPress={() => SaveOrders() }
+									>
+										<Text style={{backgroundColor:Colors.primary , borderRadius: 2,color: 'white',paddingVertical: 10 ,paddingHorizontal: 13,elevation: 5}}>Continue</Text>
+									</Pressable>
+								</View>
+							</View>
+						</View>
+					</Modal>
+
+
+					{/* {( vatStatu != 'true'  && hasVatProducts) ?
 						<View style={{ justifyContent: 'center' }} >
 							<Modal
 								animationType="slide"
 								transparent={true}
 								visible={true}
 								onRequestClose={(  ) => {
-									Alert.alert("Modal has been closed.");
+										Alert.alert("Modal has been closed.");
 								}} >
 
 								<View style={styles.centeredView}>
@@ -266,7 +414,7 @@
 						</View>
 					:
 						<View></View>
-					}
+					} */}
 
 					{(ActInd == true) ?
 						<View style={{ flex:1,position:'absolute',justifyContent:'center',height:'100%',width: '100%',backgroundColor: '#ededed',zIndex:9999,opacity: 0.5}} >
@@ -313,19 +461,29 @@
 								Object.values(data).map((value , key) => {
 									{currentSelectedLoadName = Object.keys(value)[0]}
 									return (
-										<View key={key}>
+										<View key={generateRandString()}>
 											{Object.values(value).map((val , k) => {
-
 												{currentSelectedId = val.id}
 												{valuetem = (val.order_qty).toString()}
-
-												{(selectedBuyerId != '') ? setUpdatedDataArray.push({"dnum":currentSelectedLoadName,"route":selectedRoute,"vehicle":selectedVehicle,"driver":selectedDriver,"buyer":selectedBuyerId,"sitem":currentSelectedId,"qty":val.order_qty,"credit":"NO","sale_price":val.sale_price}) : ''}
+												{setTotalAmount = (parseFloat(setTotalAmount) + parseFloat(valuetem * val.sale_price) )}
+												
+												{(selectedBuyerId != '') ? setUpdatedDataArray.push({'VATStatus' : val.VATstatus ,"dnum":currentSelectedLoadName,"route":selectedRoute,"vehicle":selectedVehicle,"driver":selectedDriver,"buyer":selectedBuyerId,"sitem":currentSelectedId,"qty":val.order_qty,"credit":"NO","sale_price":val.sale_price}) : ''}
 												return(
-													<View style={styles.mainBox} key={key}>
-														<View style={styles.itemBox} key={key}>
-															<Image source={{uri:imagePrefix+''+val.img}} style={{width: 50, height: 55, marginRight: 8}} />
-															<View key={key}>
-																<Text key={key} style={{ fontSize: 15, fontWeight: 'bold', }} allowFontScaling={false}>
+													<View style={(win.width > 500) ? styles.mainBoxTab : styles.mainBox } key={generateRandString()}>
+														<View style={styles.itemBox} key={generateRandString()}>
+															{(val.itemcategory != "EGGS") ? 	
+																<Pressable onPress={() => {  updateVATStatusOfProduct(currentSelectedLoadName ,val.id ) }}>
+																	<CheckBox
+																		checked={val.VATstatus}
+																		onPress={() => {  updateVATStatusOfProduct(currentSelectedLoadName ,val.id ) }}
+																	/>
+																</Pressable>
+															:
+																<View style={{width: 60}} ></View>
+															}
+															<Image source={{uri:imagePrefix+''+val.img}} style={( win.width > 500 ) ? {width: 50, height: 55, marginRight: 8} : {width: 40, height: 45, marginRight: 8} } />
+															<View key={generateRandString()}>
+																<Text key={generateRandString()} style={{ fontSize: 15, fontWeight: 'bold', }} allowFontScaling={false}>
 																	{((val.name.length > 20) ? (val.name).substring(0 , 20)+'..'  : val.name )}
 																</Text>
 																<Text style={{fontSize: 10}} allowFontScaling={false}> Available Stock </Text>
@@ -339,11 +497,41 @@
 															</View>
 														</View> */}
 					
-														<View key={key} style={styles.inputBox}>
-															<TextInput keyboardType="numeric" placeholder="Qty" value={valuetem} ref={(value) => {}} style={styles.textInput} onChange={(value) => { updateQty(currentSelectedLoadName ,val.id , value.nativeEvent.text) } }/>
+														<View key={generateRandString()} style={styles.inputBox}>
 
-															<TextInput keyboardType="numeric" placeholder="Price" value={val.sale_price} style={styles.textInput} onChange={(value) => { updatePrice(currentSelectedLoadName ,val.id , value.nativeEvent.text) } } />
-															<Text style={{ paddingHorizontal: 10,paddingVertical: 16,backgroundColor: '#ededed',borderWidth: 1 , borderColor: Colors.primary }}>{ (valuetem * val.sale_price).toFixed(2) }</Text>
+															<TextInput keyboardType="numeric" placeholder="Qty" defaultValue={valuetem} ref={(value) => {}} style={styles.textInput}
+															onPressIn={() => { setIsKeyboardOpen(true) }}
+															onEndEditing={(value) => { updateQty(currentSelectedLoadName ,val.id , value.nativeEvent.text) } }/>
+
+															<TextInput keyboardType="numeric" placeholder="Price" defaultValue={val.sale_price} style={styles.textInput} 
+															onPressIn={() => { setIsKeyboardOpen(true) }}
+															onEndEditing={(value) => { updatePrice(currentSelectedLoadName ,val.id , value.nativeEvent.text) } } />
+
+															<Text style={{ minWidth:70,paddingHorizontal: 10,paddingVertical: 15,backgroundColor: '#ededed',borderWidth: 1 , borderColor: Colors.primary }}>{ (valuetem * val.sale_price).toFixed(2) }</Text>
+
+															<Text style={{ minWidth:40,paddingHorizontal: 10,paddingVertical: 15,backgroundColor: '#ededed',borderWidth: 1 , borderColor: Colors.primary }}>
+																{( val.VATstatus == true )?
+																	<View>
+																		<Text>
+																			{( ((valuetem * val.sale_price) *1.20) - (valuetem * val.sale_price) ).toFixed(2)}
+																		</Text>
+																	</View>
+																:
+																	<Text>0</Text>
+																}
+															</Text>
+															<Text style={{ minWidth:40,paddingHorizontal: 10,paddingVertical: 15,backgroundColor: '#ededed',borderWidth: 1 , borderColor: Colors.primary }}>
+																{( val.VATstatus == true )?
+																	<View>
+																		<Text>
+																			{(  (((valuetem * val.sale_price) *1.20) - (valuetem * val.sale_price)) + (valuetem * val.sale_price) ).toFixed(2)}
+																		</Text>
+																	</View>
+																:
+																	<Text>{ (valuetem * val.sale_price).toFixed(2)}</Text>
+																}
+															</Text>
+
 														</View>
 													</View>
 												)
@@ -360,13 +548,15 @@
 						</ScrollView>
 					</View>
 					<View style={{borderTopColor: 'lightgrey' , borderTopWidth: 1}}>
-						<Pressable style={{padding: 16,backgroundColor:Colors.primary,flexDirection: 'row',justifyContent: 'center'}}><Text style={{textAlign: 'center',color: 'white',fontSize: 20}} onPress={() => { SaveOrders() }}>Proceed to place order</Text>
+						
 						{(saveOrderActivIndictor == true) ?
-							<Text ><ActivityIndicator size="small" color="white"></ActivityIndicator></Text>
+							<View style={{backgroundColor: Colors.primary,textAlign: 'center',width: '100%'}}>
+								<Text style={{textAlign: 'center'}}><ActivityIndicator size="large" color="white"></ActivityIndicator></Text>
+							</View>
 						:
-							<View></View>
+							<Pressable style={{padding: 16,backgroundColor:Colors.primary,flexDirection: 'row',justifyContent: 'center'}}><Text style={{textAlign: 'center',color: 'white',fontSize: 20}} onPress={() => { showConfirmationModel() }}> Place order and print invoice £{(parseFloat(MyTotalPrice) ).toFixed(2)} </Text>
+							</Pressable>
 						}
-						</Pressable>
 						{/* {setTotalAmount} */}
 					</View>
 				</View>
@@ -386,12 +576,20 @@
 		backgroundColor: Colors.redMaroon,
 		alignSelf: 'center',
 	},
-	mainBox: {
+	mainBoxTab: {
 		flex: 1,
 		flexDirection: 'row',
 		justifyContent: 'center',
 		borderColor: 'red',
 		height: 90,
+		paddingHorizontal: 5,
+	},
+	mainBox: {
+		flex: 1,
+		flexDirection: 'column',
+		justifyContent: 'center',
+		borderColor: 'red',
+		height: 130,
 		paddingHorizontal: 5,
 	},
 	itemBox: {
@@ -401,6 +599,9 @@
 		alignItems: 'center',
 		borderColor: 'blue',
 		height: 90,
+	},
+	 checkbox: {
+		alignSelf: "center",
 	},
 	buttonBox: {
 		flex: 1,
@@ -418,13 +619,22 @@
 		justifyContent: 'center',
 		alignItems: 'center',
 		borderColor: 'dodgerblue',
-		height: 90,
+		height: 120,
+		marginRight: 25
 	},
 	textInput: {
 		borderColor: Colors.purple,
 		borderWidth: 1,
-		width: 50,
-		color: '#000'
+		width: 60,
+		color: '#000',
+		textAlign: 'center'
+	},
+	textInputTab: {
+		borderColor: Colors.purple,
+		borderWidth: 1,
+		width: 60,
+		color: '#000',
+		textAlign: 'center'
 	},
 	activeStatus: {
 		backgroundColor: Colors.primary,
